Try using stick key to fix combo rollover issue
I'm using combos to capitalize individual characters. I'm using partial hold to allow me to start a sequence of capitalized letters with a combo. However, the current approach fails to capitalize any keypresses that occur before the character key from the initial macro is released. I'm seeing if pressing the shift modifier with a stick key solves this issue.